// src/enums.ts
export enum SecurityType {
  ETF,
  Stock,
}

export interface ETF {
  type: SecurityType.ETF;
  name: string;
  accumulating: boolean;
  isin: string;
}

export interface Stock {
  type: SecurityType.Stock;
  name: string;
  isin: string;
}

export type Security = ETF | Stock;